requirement.txt
pywhatkit
SpeechRecognition
transformers 

pywin32
joking
pyttsx3
gTTS
requests
googletrans
pillow
pyaudio
openai
torch